feat: Refactor header and footer into separate components for improved structure and maintainability
