+++ src/App.css
@@ -246,100 +246,6 @@ body, html, #root {
   padding: 0 1.5rem;
 }
 
-.header {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  z-index: 100;
-  padding: 1.5rem 0;
-  transition: background-color 0.3s ease, padding 0.3s ease;
-  backdrop-filter: blur(10px);
-  background-color: rgba(10, 14, 23, 0.8);
-  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
-}
-
-.header-scrolled {
-  padding: 0.8rem 0;
-  background-color: rgba(10, 14, 23, 0.95);
-  box-shadow: 0 4px 20px rgba(0, 0, 0, 0.2);
-}
-
-.header-container {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  max-width: 1200px;
-  margin: 0 auto;
-  padding: 0 1.5rem;
-}
-
-.logo {
-  font-weight: 700;
-  font-size: 1.5rem;
-  background: var(--accent-gradient);
-  -webkit-background-clip: text;
-  -webkit-text-fill-color: transparent;
-  background-clip: text;
-  cursor: pointer;
-}
-
-.nav-links {
-  display: flex;
-  gap: 2rem;
-}
-
-.nav-link {
-  color: var(--text);
-  text-decoration: none;
-  font-weight: 500;
-  transition: color 0.2s ease;
-  position: relative;
-  padding: 8px 4px;
-}
-
-.nav-link:hover, .nav-link:focus {
-  color: var(--accent);
-}
-
-.nav-link::after {
-  content: '';
-  position: absolute;
-  bottom: -6px;
-  left: 0;
-  width: 100%;
-  height: 2px;
-  background: var(--accent-gradient);
-  transform: scaleX(0);
-  transform-origin: right;
-  transition: transform 0.3s ease;
-}
-
-.nav-link:hover::after, .nav-link:focus::after {
-  transform: scaleX(1);
-  transform-origin: left;
-}
-
-/* Mobile menu button */
-.mobile-menu-btn {
-  display: none;
-  background: transparent;
-  border: none;
-  cursor: pointer;
-  padding: 10px;
-  z-index: 101;
-  position: relative;
-}
-
-.menu-line {
-  display: block;
-  width: 26px;
-  height: 2px;
-  background: var(--text);
-  margin: 5px 0;
-  transition: all 0.3s ease;
-}
-
 .hero-content {
   display: flex;
   flex-direction: column;
@@ -539,7 +445,7 @@ body, html, #root {
   height: 14px;
   transform: translateX(-6px);
   border-radius: 50%;
-  background: var(--accent);
+  background: var (--accent);
   z-index: 1;
 }
 
@@ -865,19 +771,6 @@ body, html, #root {
   color: var(--accent);
 }
 
-.footer {
-  background: rgba(10, 14, 23, 0.95);
-  padding: 2rem 0;
-  text-align: center;
-  margin-top: 0;
-  border-top: 1px solid rgba(255, 255, 255, 0.05);
-}
-
-.footer-content {
-  color: var(--text-secondary);
-  font-size: 0.9rem;
-}
-
 @media (max-width: 768px) {
   .hero-title {
     font-size: 2.5rem;
@@ -898,49 +791,6 @@ body, html, #root {
   
   .btn {
     width: 100%;
-  }
-
-  .nav-links {
-    display: none;
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100vh;
-    background: rgba(10, 14, 23, 0.98);
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    z-index: 100;
-    gap: 2rem;
-    transition: all 0.3s ease;
-  }
-  
-  .nav-links.active {
-    display: flex;
-  }
-  
-  .mobile-menu-btn {
-    display: block;
-  }
-  
-  /* Mobile menu animation */
-  .mobile-menu-btn.active .menu-line:nth-child(1) {
-    transform: rotate(45deg) translate(5px, 5px);
-  }
-  
-  .mobile-menu-btn.active .menu-line:nth-child(2) {
-    opacity: 0;
-  }
-  
-  .mobile-menu-btn.active .menu-line:nth-child(3) {
-    transform: rotate(-45deg) translate(5px, -5px);
-  }
-  
-  /* Improve touch targets */
-  .nav-link {
-    font-size: 1.2rem;
-    padding: 12px;
   }
   
   /* Hero section mobile spacing */
