+++ src/components/layout/Footer.css
@@ -1,0 +1,18 @@
+.footer {
+  background: rgba(10, 14, 23, 0.95);
+  padding: 2rem 0;
+  text-align: center;
+  margin-top: 0;
+  border-top: 1px solid rgba(255, 255, 255, 0.05);
+}
+
+.footer-content {
+  color: var(--text-secondary);
+  font-size: 0.9rem;
+}
+
+@media (max-width: 768px) {
+  .footer-content {
+    padding: 0 1rem;
+  }
+}
